requirements.txt
azure.core.credentials
azure.ai.formrecognizer